phy/utils: move DFIRateConverter to dfi.py and clean it up
